feat: GET updated users info after POST
