+++ src/apiCalls.js
@@ -3,7 +3,7 @@ let recipeData, ingredientsData, usersData, user
 import {showRecipesPage, displayRecipes} from './domUpdates.js'
 import { selectRandomUser } from '../src/RecipeRepository.js';
 
-const usersResponse = fetch('https://what-s-cookin-starter-kit.herokuapp.com/api/v1/users').then((response) => response.json());
+const usersResponse = fetch('http://localhost:3001/api/v1/users').then((response) => response.json());
 const recipeResponse = fetch('https://what-s-cookin-starter-kit.herokuapp.com/api/v1/recipes').then(response => response.json());
 const ingredientsResponse = fetch('https://what-s-cookin-starter-kit.herokuapp.com/api/v1/ingredients').then((response) => response.json());
 
@@ -16,13 +16,43 @@ ingredientsData = ingredients.ingredients
 
 showRecipesPage();
 user = selectRandomUser(usersData)
+console.log('user before:', user)
 displayRecipes(recipeData)
 return user
 })})
+
+const saveRecipe = (userData, recipeData) => {
+  fetch('http://localhost:3001/api/v1/usersRecipes', {
+    method: 'POST',
+    body: JSON.stringify({
+      userID: userData.id,
+      recipeID: recipeData.id
+    }),
+    headers: {
+      'Content-Type': 'application/json'
+    }
+  })
+  .then(response => response.json())
+  .then(getUpdatedUsers(userData.id))
+  .catch(err => alert(err));
+};
+
+const getUpdatedUsers = (id) => {
+  fetch('http://localhost:3001/api/v1/users')
+  .then(response => response.json())
+  .then(users => {
+    usersData = users.users
+    user = usersData.find(user => user.id === id)
+    return user
+  })
+  .catch(err => alert(err))
+}
 
 export {
     recipeData,
     usersData,
     ingredientsData,
-    user
+    user,
+    saveRecipe,
+    getUpdatedUsers
 }
